requirements.txt
aiohttp==1.0.5
async-timeout==3.0.1
attrs==21.2.0
cffi==1.14.5
chardet==4.0.0
discord-buttons==0.5
discord-command==0.0.2
discord-components==1.0.0
discord.py==1.7.3
DiscordUtils==1.3.4
ffmpeg==1.4
ffmpeg-media-converter==0.2.1
idna==3.2
multidict==5.1.0
numpy==1.20.3
pafy==0.5.5
pycparser==2.20
pygame==2.0.1
pynacl==1.4.0
six==1.16.0
typing-extensions==3.10.0.0
wavelink==0.9.9
websockets==3.4
yarl==1.6.3
youtube-dl==2021.6.6
